+++ src/main/java/org/algohub/engine/judge/CsharpJudge.java
@@ -3,6 +3,7 @@ package org.algohub.engine.judge;
 import org.algohub.engine.pojo.JudgeResult;
 import org.algohub.engine.pojo.Problem;
 import org.algohub.engine.pojo.Function;
+import org.algohub.engine.bo.StatusCode;
 
 /**
  * C# language online judge.
@@ -15,6 +16,6 @@ public class CsharpJudge implements JudgeInterface {
    */
   public JudgeResult judge(final Function function, final Problem.TestCase[] testCases,
       final String userCode) {
-    return new JudgeResult(StatusCode.ACCEPTED.toInt(), null, null, null, null, 0, 0, 0, 0);
+    return new JudgeResult(StatusCode.ACCEPTED, null, null, null, null, 0, 0, 0, 0);
   }
 }
