+++ src/main/java/org/algohub/engine/judge/RubyJudge.java
@@ -12,6 +12,7 @@ import org.algohub.engine.util.ObjectMapperInstance;
 
 import java.io.File;
 import java.io.IOException;
+import org.algohub.engine.bo.StatusCode;
 
 
 public class RubyJudge implements JudgeInterface {
@@ -51,21 +52,21 @@ public class RubyJudge implements JudgeInterface {
           break;
         }
         case 1: { // Compile Error
-          result = new JudgeResult(StatusCode.COMPILE_ERROR.toInt(),
+          result = new JudgeResult(StatusCode.COMPILE_ERROR,
               createFriendlyMessage(runResult.getStderr()), null, null, null, 0, testCases.length,
               0, 0);
           break;
         }
         default: { // Other runtime errors
           result =
-              new JudgeResult(StatusCode.RUNTIME_ERROR.toInt(), runResult.getStderr(), null, null,
+              new JudgeResult(StatusCode.RUNTIME_ERROR, runResult.getStderr(), null, null,
                   null, 0, testCases.length, time, 0);
         }
       }
       JudgeInterface.removeRecursive(tmpDir.toPath());
       return result;
     } catch (IOException | InterruptedException ex) {
-      return new JudgeResult(StatusCode.RUNTIME_ERROR.toInt(),
+      return new JudgeResult(StatusCode.RUNTIME_ERROR,
           ex.getClass().getName() + ", " + ex.getMessage(), null, null, null, 0, testCases.length,
           0, 0);
     }
